Add force-less umount
Bug description:
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1402834

Summary:
`lxc-destroy` uses umount with --force flag in order to unmount
shared/binded folders. Due to the bug described above it causes
to unmount the folder from the host machine aswell.

This is a compiled version of the following file from version v2.28:
https://github.com/karelzak/util-linux/blob/master/sys-utils/umount.c

The compiled version has the following line removed:
https://github.com/karelzak/util-linux/blob/master/sys-utils/umount.c#L579